backend/requirements.txt
fastapi
uvicorn[standard]
requests
apscheduler
python-dotenv
supabase
pydantic
